feat(home): Create component
