+++ next-frontend/styles/index.module.scss
@@ -1,0 +1,10 @@
+.root {
+  overflow: hidden;
+  width: 100%;
+}
+
+.bananaNav {
+  padding-bottom: 20%;
+  padding-left: 33%;
+  width: 100%;
+}
